Refactored com.api.pojo to com.api.request.model
