+++ src/test/java/com/api/request/model/CreateJobPojo.java
@@ -1,4 +1,4 @@
-package com.api.pojos;
+package com.api.request.model;
 
 import java.util.List;
